requirements.txt
gitpython
prompt_toolkit
coolname
pygments
openai
openai[datalib]
tenacity
flask
anthropic
torch
transformers
sentence_transformers==2.2.2
InstructorEmbedding